pyproject.toml
[project]
name = "nonebot-plugin-pingti"
dynamic = ["version"]
description = "Shop with lower prices by searching with alternative keywords."
authors = [{ name = "student_2333", email = "lgc2333@126.com" }]
dependencies = [
    "nonebot2>=2.4.0",
    "nonebot-plugin-alconna>=0.54.1",
    "nonebot-plugin-waiter>=0.8.0",
    "cookit[loguru,nonebot-alconna,pydantic]>=0.8.3",
    "httpx>=0.27.0",
]
requires-python = ">=3.9,<4.0"
readme = "README.md"
license = { text = "MIT" }

[project.urls]
homepage = "https://github.com/lgc-NB2Dev/nonebot-plugin-pingti"

[tool.pdm.build]
includes = []

[tool.pdm.version]
source = "file"
path = "nonebot_plugin_pingti/__init__.py"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
